mwaa/requirements.txt
smart-open==6.2.0
pydantic==1.9.1
airflow-provider-great-expectations==0.1.1